Реализует отрисовку страницы через контроллер
